Restore single AbstractTxView, move TxView to separate file
